fixed bug in char literals

+++ src/codegen/llvm_codegen.cpp
@@ -204,7 +204,7 @@ Opt<llvm::Value*> Codegen::codegen_value(ASTNode& node) {
                 case PrimaryExpr::void_:
                     return llvm::Constant::getNullValue(ir_builder->getPtrTy());
                 case PrimaryExpr::char_literal:
-                    return ir_builder->getInt32(std::get<int64_t>(prim.value));
+                    return ir_builder->getInt8(std::get<int64_t>(prim.value));
                 default:
                     internal_panic("codegen not implemented for '{}'", node.name());
             }
